+++ modules/authentication/server-scala/src/main/scala/session/model/Session.scala
